Ajouté optimisation des images, corrigé dépassement de la taille des éléments et corrigé bug de clic à travers la fenêtre modale (#14)
* Amélioration visuelle de l'interface utilisateur

* Delete calendar-app/public/images/RALLYE APPART_RÉMAA_2025.png

* retiré index.txt

* Amélioration visuelle de la version mobile

* Ajouté mots de passes ADMIN

* Ajouté de la robustesse à l'initialisation du calendrier pour éviter une erreure fatale en cas de malformation des données.

* Clarifié la nature des boutons filtres

* ajouté '?=v5' aux fichiers .css

* caché les fenêtres modales par défaut

* Ajouté 'assoc-colors.json' au .gitignore

* Corrigé bug de téléchargement des images

* Create motsDePasses.json

* Mis à jour le README

* Retiré nécessité d'ajouter le tag aux images dans le README

* Refactorisé code de téléchargement des images et ajouté support pour les liens d'inscriptions

* Amélioré imgworker pour produire différentes tailles d'images

* Ajouté optimisation des images .png, corrigé le dépassement du calendrier des bords de l'écran et corrigé bug de click à travers la fenêtre contextuelle

* Ajouté optimisation des images .jpg, .webp et .gif et corrigé comportement de changement de taille des images

* Corrigé problèmes de mise à l'échelle

---------

+++ converter/import_events.py
@@ -72,21 +72,28 @@ def save_event_attachments(vevent: Event, assoc: str, utilisateur: str) -> Path:
         if not nom_fichier:
             break
         
+        if ( SORTIE_IMAGES / assoc / ( nom_fichier.replace(".","_") ) ).exists():
+            return SORTIE_IMAGES / assoc / nom_fichier.replace(".","_") / "tailles.txt"
+        
         répertoire_source = RACINE_NEXTCLOUD / "data" / utilisateur / "files/Calendar" / nom_fichier
         répertoire_sortie = SORTIE_IMAGES / assoc / nom_fichier
         répertoire_sortie.parent.mkdir(parents=True, exist_ok=True)
+        os.chmod(répertoire_sortie.parent.parent, 0o777)
+        os.chmod(répertoire_sortie.parent, 0o777)
         
         try :
             if répertoire_sortie.exists():
-                    if répertoire_source.stat().st_size != répertoire_sortie.stat().st_size or hash_file(répertoire_source) != hash_file(répertoire_sortie) :
-                        shutil.copy2(répertoire_source, répertoire_sortie)
+                if répertoire_source.stat().st_size != répertoire_sortie.stat().st_size or hash_file(répertoire_source) != hash_file(répertoire_sortie) :
+                    shutil.copy2(répertoire_source, répertoire_sortie)
+                    os.chmod(répertoire_sortie, 0o777)
             else :
                 shutil.copy2(répertoire_source, répertoire_sortie)
+                os.chmod(répertoire_sortie, 0o777)
         except Exception:
             pass
 
         return répertoire_sortie
-    pass
+    return None
 
 def main():
 
@@ -179,4 +186,3 @@ def main():
 
 if __name__ == "__main__":
     raise SystemExit(main())
-
